Fix regtest block generation.
The "setgenerate" RPC method on regtest just calls createNewBlock
directly to generate blocks on demand.  Thus when a block is found,
we should just return from the method, rather than throw an exception
(which might be appropriate for a separate mining thread, but here
just kills the RPC handler thread).