Fix use-after-free in mempool tx collector.
AddTransactionsToBlockIfPossible keeps a local map of transactions
to dependent other transactions in dependentTransactions.  Those
are filled in from PrioritizeMempoolTransactions.

However, the underlying COrphan instances that are referenced from
the map were (indirectly) allocated on the stack only in
PrioritizeMempoolTransactions, so that they would be gone and used
after being freed from e.g. PrioritizeTransactions.