Always send inv for requested blocks.
This is an alternative (and cleaner) fix for the issue from
b18a72a7b63bdd023843d9e5852a762deca34991.

We bypass PushInventory and the filter for already-sent inv's when
replying to a getblock request.  This ensures that the peer who asks
for the chain during syncing will always receive all required blocks
in order, and will be able to sync to the chain properly.

This may cause some "redundant" inv messages to be sent, but that will
be of minor impact as it only applies to explicitly asked for inv's and
the messages themselves are rather small as well (in particular, they are
just *block hashes* and not full blocks!).