Simplify logic in CheckProofOfStake.
When checking PoS, one of the conditions is that all inputs for
the coin-stake transaction should be for the same address / script.
The code verifying this condition was quite intransparent, though,
as it used a lambda-expression to count the inputs matching the script
of the first input, and then verified that the count is the total number
of inputs.

This change simplifies the code, making it easier to read and understand.
Instead of the counting, we simply iterate over all inputs, and explicitly
verify the condition.

As a side-effect, this also makes the verification a bit faster, as we
a) avoid loading and checking the first input twice, and b) exit early
if we already found a mismatch.