Fix flaky wallet regtests.
Fix wallet.py and walletbackup.py, which were flaky before and had
random failures due to missing syncs.  At the same time, we can also
add proper accounting of fees to walletbackup.py, instead of the
previous "guessing" which was not fully accurate and failed as well
sometimes due to the tolerance being not large enough.