Avoid selecting too small outputs for staking.
When computing the weight of a staked coin, the "minimum" that will
be allowed is 400 DIVI seconds (everything else gets truncated to
exact zero and yields an impossible hash target).

On regtest where the spacing between blocks might only be a few
seconds (and there is no minimum age for staking), we have to make
sure that coins are thus at least 400 DIVI, or else we may select
ones that then fail to produce a block.