Use GetTime() for mn sync process.
Instead of using ticks in CMasternodeSync::Process and assuming the
function is called every second, use GetTime() compared to a known
timestamp of the last sync step to trigger the individual stages.

This is much clearer to understand, more explicit, and it also enables
using mocktime more easily for regtests.

Also includes some general code cleanups:  Two constants that are used
only internally in masternode-sync.cpp are moved to the cpp file to
make this explicit, and since we have a class, use a member variable
instead of a function-static one to hold the last call timestamp.