Simplify masternode rank for unknown nodes.
If the asked-for masternode is unknown or not active itself, return
always -1 from GetMasternodeRank, and handle that as "rank too high
to win" in the callers.  That's equivalent to returning an index
beyond the list of masternodes anyway (assuming we have more than 10
in our list).

The check for "we have the given block" is also redundant, as all
callers only invoke the method for "current tip - 100 blocks", which
is of course a block that is known.