Only check rank up to a given maximum.
When computing the masternode ranks, we only really need to know if
the rank is in the top-10 or 20, but not the actual rank beyond that
(if the node is not).

Explicitly pass that information to GetMasternodeRank.  By doing so,
we can exit the search early; that is, however, just a marginal benefit,
as the method is mostly called if the node actually is in the list
of winners.  But by having this restriction, we can more easily add
in caching of the top ranks in a follow-up step.