Disable masternode scheduled check on regtest.
Usually, masternodes are only considered as payment winner if they
are not already receiving another payout in the next eight blocks.
On regtest, this is annoying -- because when testing with a very small
number of masternodes, it means that those will always be scheduled and
also skipped because of that check.  (And worse, the rule can even force
particular non-random patterns in the payouts.)

This change disables the scheduled check on regtest, so it is easier
to test things like the payout schedule / chances.