Pass explicit seed hash to CalculateScore.
Instead of calculating the score of a masternode based on a block
height, pass the underlying block hash (which is used to seed the
score randomness) directly.  This makes the code more explicit, and
will allow better separation of the translation from height to hash
(which is somewhat prone to reorg issues).

Also moves the -100 block height offset into GetBlockHashForScoring
(GetMnBlockHash renamed), so that this is only specified in a single
place.  The function itself is now passed the block height for which
we want the winner / score (not the offset height).