Store scoring hash with CMasternodePaymentWinner.
In each CMasternodePaymentWinner message, store the scoring hash
in addition to the block height.  This is what ultimately determines
the winner, and is thus the "proper key" for those votes (while the
height is prone to reorgs).

Since the wire protocol serialises (and hashes) these with height and
we do not want to change the protocol, we only compute the seed hash
internally after parsing a new message, and then use it for validating
against our masternode scoring.  (And in the future, we will use it
also to index our in-memory caches.)