Compute scoring hash from CBlockIndex.
Instead of passing a block height and using chainActive, pass a
CBlockIndex and search in its explicit ancestor chain when computing
the scoring hash to be used for a block.  This further decouples the
masternode logic from chainActive as a global and makes it more robust
against reorgs.

The only remaining instance when we still use chainActive is for
computing the seed hash after parsing a CMasternodePaymentWinner
message from the wire.  In this case, we only have a block height,
so using chainActive as context is the best we can do (unless we
would want to change the wire protocol itself).