Track payment votes more explicitly.
Instead of tracking votes for payment winners "indirectly" through
the height that each masternode has voted for last, just explicitly
keep track of the votes in the block payee data structure as a set
of masternodes who did vote.  (There will be only 10 anyway, so it
does not use up a lot of memory.)

This is simpler, more robust, and fixes potentially weird behaviour
when voting messages are received out-of-order (i.e. not in increasing
order of block heights).