Extra check for staking vault coinstakes.
Implement a new function CheckCoinstakeForVaults, which does the extra
checks for vault staking (i.e. that the vault owner gets paid back
their due in a single output to the same script).

For now it has unit tests but is not yet invoked in the actual validation
logic; this will be added as a softfork with block-version activation
(a la BIP30) later.