Make data pushes in CombineSignatures minimal.
CScript's operator<< for pushing data is not necessarily doing it in
the minimal way (namely when a single-byte op-code like OP_5 or
OP_1NEGATE can be used).  This means that CombineSignature (e.g. from
using signrawtransaction) may produce signatures that fail verification
with SCRIPT_VERIFY_MINIMALDATA.  For "normal" signature types like
P2PKH this does not happen; but the OP_0 and OP_1 in the signature
of vault scripts trigger the bug.

This fixes CombineSignatures to push data in a minimal way.  It also
adds a unit test that SignSignatures + CombineSignatures works for
signing vault scripts both as vault and owner.