Extend createrawtransaction RPC method.
Extend the createrawtransaction RPC method, with two main changes:  First,
the "outputs" argument can now be either a JSON object (as before) or
a JSON array.  In the latter case, the elements must be single-element
objects that specify the outputs.  This matches modern Bitcoin Core,
and allows more explicit control over the order of outputs.

Second, if an "address" passed for the outputs is actually a hex string,
interpret it as literal script rather than trying to parse the address.
This allows more flexible transaction creation, e.g. for staking vaults.

The change is backwards compatible.  Any valid usage of createrawtransaction
stays valid (but some previously invalid calls are now valid also).