Use separate queues to manage block mining & sync before compare
Issues with the reliability of this test stem mainly
form a lack of synchronization with the nodeId=1 with
the longest chain resulting in an rpc error when trying
to produce a new block. By synchronizing before comparing
the issue is resolved. Further, using different threads
for mining blocks as well as allowing them to fail
allows for a cleaner comparison to a real world setting