Ensure wallet is unlocked for masternode RPCs.
Some of the masternode RPCs (those related to an owned masternode)
require private keys from the local wallet.  For them, explicitly
ensure the wallet is unlocked before proceeding.

Without this, the error message received by the user when the wallet
is unlocked is confusing; with this change, they will get a clear
message that the wallet needs to be unlocked (like they already do for
e.g. sendtoaddress).