Avoid looking up masternode by payee.
Looking up a masternode by payee script is error-prone, as it is
not a unique identifier.  Multiple masternodes could be using the
same collateral address (and may in the future use the same reward
address).

The only place where this is being used is in verifying a mnw
message.  Instead of the previous logic relying on the lookup
(which was hard to read anyway), just explicitly check that the
payee matches one of the masternodes in our payment queue.