Unify (broadcast)startmasternode code.
startmasternode and broadcaststartmasternode do something very similar,
namely processing and broadcasting the activation message for a masternode
(typically one owned by the person owning the node the RPC is run on as
well, but that's not specially required).

However, both routines used different code, and in particular
broadcaststartmasternode had a bug that didn't insert the broadcast
into mapSeenMasternodeBroadcast.  Thus the node would send out an
announcement of the broadcast message, but wouldn't be able to reply
to peers requesting the actual data (unless startmasternode with "defer"
and the subsequent broadcaststartmasternode would be done on the
same node).

This commit reduces code duplication by using a single code path to
process the masternode broadcast for both RPC methods (with the difference
that startmasternode first constructs the broadcast itself).  It also
extends the mnoperation.py regtest to use a deferred broadcast relay
for one of the two nodes to test it.