Consolidate masternode input age logic.
This simplifies and consolidates the logic used to verify confirmations
and age of the collateral UTXO for masternodes.  Instead of various
functions we now have just one main function, which looks up the block
index when the UTXO was confirmed.

With this, we can check that it actually has the required confirmations,
that the sigtime matches (is later than the block when it got the
final confirmation necessary), and also cache the block hash for
later input age computation.