Tweak sigTime as necessary for delayed broadcast.
When creating a delayed masternode broadcast, check the sigtime against
the collateral confirmation block and make sure that it meets the
network criterion (signature must not be earlier than the timestamp
of the block that confirmed the collateral 15 times).

If the condition is violated, just use "now + one hour" as sigtime.