Do not try to register a deferred masternode.
When startmasternode is called with defer set to true, do not even
attempt to register the masternode; just create the broadcast data
and return it.

This makes it possible to use the function even when the masternode
is not yet valid, e.g. because the collateral does not have sufficient
confirmations yet.