Regtest for proposed mn vaults.
This adds a new regtest mnvaults.py, which manually implements the
proposal for masternode vaults with a temporary key and a pre-signed
unvault transaction.

It sets up a masternode with a collateral at a temporary address,
for which we need not retain the private key.  Instead we just keep
a pre-signed transaction to withdraw the funds to a specified address
and the signed masternode broadcast in storage, and use them to
first run the masternode and then unvault the funds.