Add masternode directly in ProcessBroadcast.
Instead of having a function CMasternodeBroadcast::CheckInputsAndAdd,
we now just have a read-only "CheckInputs" method.  The updates to
the state (in particular, adding the masternode to our list / removing
an existing other one) are done in CMasternodeMan::ProcessBroadcast,
where they belong together with some other updates we do anyway.

This also allows us to ensure that we add the broadcast message to
mapSeenMasternodeBroadcast before we actually relay the inv.