Verify ping contained in mnb message.
This adds logic to verify the lastPing contained in mnb messages
when processing those.  Without it the ping itself could be malleated
by an attacker, and we would then happily save it in mapSeenMasternodePing
in the wrong version (and relay it from there).

To this end, we also refactor CMasternodePing::CheckAndUpdate to make
it independent of global state.  Instead of looking up and updating the
masternode from the global masternode manager, it is passed in so that
the method itself is not touching global state.  This allows us to
purely verify the ping when processing a broadcast (updating a temporary
CMasternode instance) and before changing any global state.