Implement ActivationState class.
This adds the new ActivationState utility class.  It is used to
abstract fork activation away from other parts of the code.  An
instance is based on a CBlockIndex representing the chain tip being
validated, and it can be queried for the state of various forks
with IsActive.

Activation happens by block timestamp, as this is totally straight-forward
and more or less the simplest form of activation possible.  This also
makes it independent of the block context (unlike e.g. median time)
and allows us to theoretically use fork activation even for things like
changing the serialisation format of blocks.

For now, no fork is defined apart from a dummy one used in testing.
In the future, we will define real forks with this framework (e.g. the
staking-vault changes).