New regtest for OP_META.
This adds op_meta.py, a new regtest for some basic features of
OP_META transactions.  In particular, it verifies the size limits
for when OP_META transactions are and are not standard, that
multiple OP_META outputs are not standard ever, and empirically
figures out the fee required to send an OP_META transaction with
small and large payload.