Introduce bare txid for transactions.
This defines a new method on CTransaction, which computes the "bare txid".
This is a hash that commits to the main data of the transaction, but does
not take any signatures into account.

It corresponds to the txid of a segwit transaction in Bitcoin (while the
normal txid would correspond to the wtxid of such a transaction).