Bare txid for GetTransaction / GetWalletTx.
This extends the two functions GetWalletTx and GetTransaction to (also)
find transactions by bare txid, not just by the normal txid.

These methods are mainly used in places where we need to look up e.g.
the previous transaction to a spend, so that we can know the address
that is being spent or the value of the input.

The change is fine to do (won't cause any extra consensus changes)
because all it does is make those methods return the correct previous
transaction (for after the fork) in cases where they would have
failed otherwise (since both are SHA-256d hashes and thus cannot have
collisions).  Also actual checks that some spent coin actually exists
are the explicitly on the consensus-level anyway.