FakeWallet for testing.
wallet_coinmanagement_tests includes a wallet with a real seed and
usable for testing.  This is useful also for other tests; this commit
splits out the logic into a reusable FakeWallet, and updates the test
to use that instead.

It also makes the code use FakeBlockIndexWithHashes, instead of just
the FakeBlockIndexChain with a custom block map.