Reduce cs_main lock in SendMessages.
In SendMessages, we were locking cs_main almost the entire time.
We needed the lock for the address rebroadcast, which has been moved
out already, and the State() function accessing a node's state.

This changes the code to not lock cs_main the entire time, but instead
only briefly to get the node state, and then we work with a local
reference to the state without the main lock held.