Remove CMainCleanup.
CMainCleanup was used to free some memory on process exit; but this
is unnecessary, since the OS will reclaim that memory anyway, and
there are no other side effects we need of the destructors.

Removing this also solves an issue with undefined behaviour for the
destruction order of e.g. mapOrphanTransactions and the CMainCleanup
instance itself.