Use CPeer's serialiser for mnp/mnb messages.
When broadcasting mnb or mnp messages, use the serialisation stream
in CPeer instead of a custom serialiser.  This makes the code simpler
(since that is what the stream in CPeer is for), but also makes sure
that the proper protocol version is used based on the peer, not the
default version we support.