Return normal hash as bare txid for coinbase.
For coinbase transactions, we have to return the normal txid as the
"bare" hash.  For them, there is not a real signature that could be
malleated anyway, but instead the scriptSig is needed to make sure
each coinbase transaction has a different hash from any others.

In practice, this has no relevance, since no new coinbase transactions
from PoW are created anyway after the future mainnet fork activation of
segwit light.  But for regtests it is needed (and in general required to
make sure all is correct).