Force assertion when zmq message matches block hash
Since the node that is publishing zmq messages is the same one
mining a block, there are two messages that come out from that
single event: one tx from mining rewards and one from block
generation. When these messages arrive out of order the assertion
in the test can fail.