Reset timeout for mn sync after blockchain sync.
When syncing masternodes, we record for each class of items when we
started syncing that class, and fail the sync if we have not received
any replies after some timeout.

However, when the blockchain is not yet synced, we "pause" the masternode
sync on purpose until it is up-to-date.  But this leads to spurious timeouts
(and thus mn sync failures) as the start time for syncing masternode list
items will be before the blockchain sync and will have timed out while
waiting for the blockchain.

This change fixes the issue by resetting the masternode list start
time when the blockchain is caught up.