Set IsBlockchainSynced time horizon to 10 minutes.
IsBlockchainSynced from MasternodeHelpers.cpp is used to determine
if the blockchain is "likely" up-to-date or still syncing.  The test
used, however, is very relaxed, and it is considered up-to-date even
when the last block is more than a day old.

This leads to issues, e.g. when we consider the blockchain synced,
resume masternode sync, but then receive mn broadcasts whose pings
refer to recent blocks we don't have yet, and thus consider them
invalid.

If this happens, it may fail masternode sync.  It can, however,
even be used as an attack vector:  An attacker could send us their
own masternode broadcasts with "old" pings that we will accept,
while all the rest of the network seems invalid.  This way, the attacker
can trick us into believing just their few masternodes are all active
ones, which may further lead to a fork in the blockchain consensus.

In this change, we drastically reduce the time horizon accepted for
a "synced" blockchain to only 10 minutes.