Add bare txid to transaction index.
When the transaction index is enabled (-txindex), this change also keeps
an index from bare txid to the disk position, so that lookups with the
index work on both normal txids and bare txids.

With this, the previous change made to GetTransaction will be improved
for enabled -txindex, and it will also allow us to look up any transaction
by bare txid, even if the fallbacks in GetTransaction fail.