Automatic conversion of some types in logging.
This adds a framework that allows us to automatically convert some types
to std::string when logging, without having to explicitly invoke
ToString() or GetHex() on them for each log call.

In this commit, we only add the functionality itself but not yet make
use of it.  This will come as follow-ups.