Fix log category filtering.
There was a bug in the log category filtering, which prevented
"all" (or just -debug) from working correctly:  The blanko check for
this case was only executed when the thread-local category set was
not yet constructed (i.e. on the first call in each thread), so that
log messages were filtered out for all subsequent calls when the option
was set to all.

This commit fixes the bug, and in general streamlines the logic around
this.  We now just construct the thread-local set if it doesn't exist, but
all further logic (including the category filtering itself) is just done
afterwards by the same code independent of whether or not the set existed
already.