New regtest NoBlocksForLongTime.py.
The new test NoBlocksForLongTime.py checks what happens when the
network does not produce blocks in a long time, in particular so that
IsBlockchainSynced returns false for restarted nodes.

In this scenario, we want to ensure that the network does not get stuck;
the test asserts that nodes that were not restarted can produce blocks,
as well as nodes restarted but with an up-to-date masternode list.

Only if a node is restarted and does not have any masternode data, it
will produce an invalid block (since it can't do the correct masternode
payment).