Handle CheckMintTotalsAndBlockPayees result.
When checking a block in ConnectBlock, actually handle the return value
of CheckMintTotalsAndBlockPayees properly.  In other words, if the block
mint is wrong, fail to connect the block in question.

This check was lost in a refactoring in
d5a84561e0e85c57f702a6fd8c119c81aa693935, but has never been released
to production (so the bug does not affect the live network).