Remove dead code in swifttx.
The code in swifttx.cpp is mostly commented out and not used at the moment.
This commit removes code that is completely dead, consisting of some
functions that are never even called from the outside (and consist of
only disabled logic as well).

This does not yet touch code that is in theory referred to e.g. in main.cpp,
which will be a follow-up step.