New "ignoreMempool" option for generateblock.
This extends the generateblock RPC call with a new option,
"ignoreMempool".  If set, the generated block will not include any
transactions from the mempool, but just the coinbase/coinstake and
any extratx listed.

With this, we can test certain situations more fine-grained, as we get
complete control over contents of a test block.