Properly update spent index in mempool.
When a transaction gets removed from the mempool, remove also its
spent-index entry.  If this is not done, the spent-index data (e.g.
as returned by getspentinfo) can get wrong if a transaction in the
mempool gets "replaced" by a double spend in a block.

A new regtest (which fails without the fix) and also unit test verify
the expected behaviour now.