Fix lock-order issue with reject messages.
When a transaction gets relayed from a peer that we actually don't accept
into the mempool, we send back a "reject" message.  This is something that
does not happen often in practice and was not covered by tests so far (but
is now with a new regtest).

If this happened, it triggered a lock-order violation:  The code processing
the "tx" message would first lock cs_main, and then a peer's cs_vSend
(for the "reject" message) while still holding onto cs_main.  On the
other hand, when relaying wallet transactions, the locks are used in the
reverse order (the cs_vSend when processing messages to send to a peer
in general, and the cs_main is used when determining whether or not a
wallet transaction is already in the main chain).

This fixes this issue, by releasing cs_main before sending the "reject"
message; there is no need to hold onto the lock then anyway.