Simplify StakableCoin.
This is a refactor of the StakableCoin utility class.  There was some
redundancy (e.g. between outputIndex and utxo), and also there was
support for "empty" StakableCoin instances, which was needed only in
one place but added extra complications to the class itself.

With this, the class and usage is simplified and easier to reason about.
It also makes it easier to switch to segwit light in the future, since
we now just pass in the UTXO outpoint from the wallet and never explicitly
deal with transaction hashes afterwards.